Separate Packets namespace from UltimaClient.
All Packets are now in UltimaXNA.UltimaPackets.
UltimaClient is now in the root UltimaXNA namespace.

This should be it for moving files. My next goal will be to refactor
UltimaClient - moving code specific to the UltimaLogin Model to the
UltimaLogin namespace, and likewise, moving code specific to the
UltimaWorld Model to the UltimaWorld namespace.

The end goal is to reduce the number of static projects scattered
throughout the project, making it easier to comprehend. Right now,
UltimaWorld relies on code from almost every namespace, and almost every
namespace has some code that calls something from UltimaWorld (the
static Entities class is a great example of this awfulness). By the end
of this refactor, UltimaWorld should be almost entirely self-contained.