Now overriding equality of Gumps, not Controls.
I placed the Equals() override for Gumps in the Controls class, which
was causing trouble when a Control of a given type (Say, ResizePic) was
disposed of; this would cause all other Controls of the same type in the
owner Control (all ResizePics, in my example) to be disposed as well.
Definitely incorrect.